+++ docs/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - document workflow framework
 - introduce EMBARGOED workflow
 - introduce workflow labels
+- ACE: Add Chromium special CVE workflow (OSIDB-5033)
+- ACE: Add Go stdlib special CVE workflow (OSIDB-5033)
 
 ### Changed
 - align existing workflows with workflow framework concepts